@@ -9,53 +9,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.0.8",
-    "@angular/cdk": "17.0.4",
-    "@angular/common": "17.0.8",
-    "@angular/compiler": "17.0.8",
-    "@angular/core": "17.0.8",
-    "@angular/forms": "17.0.8",
-    "@angular/material": "17.0.4",
-    "@angular/platform-browser": "17.0.8",
-    "@angular/platform-browser-dynamic": "17.0.8",
-    "@angular/router": "17.0.8",
+    "@angular/animations": "17.1.1",
+    "@angular/cdk": "17.1.1",
+    "@angular/common": "17.1.1",
+    "@angular/compiler": "17.1.1",
+    "@angular/core": "17.1.1",
+    "@angular/forms": "17.1.1",
+    "@angular/material": "17.1.1",
+    "@angular/platform-browser": "17.1.1",
+    "@angular/platform-browser-dynamic": "17.1.1",
+    "@angular/router": "17.1.1",
     "@heavy-duty/wallet-adapter": "^0.8.0",
     "@heavy-duty/wallet-adapter-cdk": "^0.8.0",
     "@heavy-duty/wallet-adapter-material": "^0.8.0",
     "@ngrx/component": "17.0.1",
     "@ngrx/component-store": "17.0.1",
-    "@nx/angular": "17.2.7",
+    "@nx/angular": "17.3.0",
     "@solana/spl-token": "^0.3.8",
     "@solana/wallet-adapter-base": "0.9.22",
     "@solana/web3.js": "1.87.6",
     "angularx-qrcode": "17.0.0",
-    "ngxtension": "^1.7.0",
+    "ngxtension": "^1.12.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "0.14.2"
+    "zone.js": "0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.0.8",
-    "@angular-devkit/core": "17.0.8",
-    "@angular-devkit/schematics": "17.0.8",
+    "@angular-devkit/build-angular": "17.1.1",
+    "@angular-devkit/core": "17.1.1",
+    "@angular-devkit/schematics": "17.1.1",
     "@angular-eslint/eslint-plugin": "17.0.1",
     "@angular-eslint/eslint-plugin-template": "17.0.1",
     "@angular-eslint/template-parser": "17.0.1",
-    "@angular/cli": "~17.0.0",
-    "@angular/compiler-cli": "17.0.8",
-    "@angular/language-service": "17.0.8",
-    "@nx/eslint": "17.2.7",
-    "@nx/eslint-plugin": "17.2.7",
-    "@nx/jest": "17.2.7",
-    "@nx/js": "17.2.7",
-    "@nx/workspace": "17.2.7",
-    "@schematics/angular": "17.0.8",
+    "@angular/cli": "~17.1.0",
+    "@angular/compiler-cli": "17.1.1",
+    "@angular/language-service": "17.1.1",
+    "@nx/eslint": "17.3.0",
+    "@nx/eslint-plugin": "17.3.0",
+    "@nx/jest": "17.3.0",
+    "@nx/js": "17.3.0",
+    "@nx/workspace": "17.3.0",
+    "@schematics/angular": "17.1.1",
     "@swc-node/register": "~1.6.7",
     "@swc/core": "~1.3.85",
     "@types/jest": "^29.4.0",
-    "@types/node": "18.16.9",
-    "@typescript-eslint/eslint-plugin": "6.15.0",
-    "@typescript-eslint/parser": "6.15.0",
+    "@types/node": "^18.16.9",
+    "@typescript-eslint/eslint-plugin": "6.20.0",
+    "@typescript-eslint/parser": "6.20.0",
     "assert": "^2.1.0",
     "autoprefixer": "^10.4.0",
     "crypto-browserify": "^3.12.0",
@@ -65,14 +65,14 @@
     "jest-environment-jsdom": "^29.4.1",
     "jest-preset-angular": "13.1.4",
     "jsonc-eslint-parser": "^2.1.0",
-    "nx": "17.2.7",
+    "nx": "17.3.0",
     "postcss": "^8.4.5",
     "prettier": "^3.1.0",
     "stream-browserify": "^3.0.0",
     "tailwindcss": "^3.0.2",
     "ts-jest": "^29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "5.2.2",
+    "typescript": "5.3.3",
     "util": "^0.12.5"
   }
 }
