@@ -1,29 +1,14 @@
-import { ClipboardModule } from '@angular/cdk/clipboard';
+import { CdkCopyToClipboard } from '@angular/cdk/clipboard';
 import { DecimalPipe } from '@angular/common';
-import {
-  Component,
-  EventEmitter,
-  HostBinding,
-  Input,
-  Output,
-} from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
+import { Component, EventEmitter, Output, input } from '@angular/core';
+import { MatButton } from '@angular/material/button';
+import { MatCard } from '@angular/material/card';
+import { MatIcon } from '@angular/material/icon';
 import { HdObscureAddressPipe } from '@heavy-duty/wallet-adapter-cdk';
+import { LetDirective } from '@ngrx/component';
 import { ToUserValuePipe } from '../shared';
 
 @Component({
-  standalone: true,
-  imports: [
-    DecimalPipe,
-    MatButtonModule,
-    MatCardModule,
-    MatIconModule,
-    ClipboardModule,
-    HdObscureAddressPipe,
-    ToUserValuePipe,
-  ],
   selector: 'my-bank-payment-section',
   template: `
     <mat-card class="px-4 py-8 w-[500px] h-full">
@@ -33,7 +18,7 @@ import { ToUserValuePipe } from '../shared';
 
       <div>
         <p class="font-bold text-lg">Details</p>
-        <p class="pl-4 mb-4">
+        <p class="pl-4 mb-4" *ngrxLet="memo() as memo">
           @if (memo !== null) {
             {{ memo }}
           } @else {
@@ -41,7 +26,10 @@ import { ToUserValuePipe } from '../shared';
           }
         </p>
         <p class="font-bold text-lg">Requested by</p>
-        <p class="pl-4 flex items-center gap-4 mb-4">
+        <p
+          class="pl-4 flex items-center gap-4 mb-4"
+          *ngrxLet="requester() as requester"
+        >
           @if (requester !== null) {
             {{ requester | hdObscureAddress }}
             <button
@@ -58,7 +46,7 @@ import { ToUserValuePipe } from '../shared';
         <hr class="mb-4" />
         <div class="flex justify-center items-center gap-2 mb-4">
           <img src="assets/usdc-logo.png" class="w-12 h-12" />
-          <p class="text-4xl">
+          <p class="text-4xl" *ngrxLet="amount() as amount">
             @if (amount !== null) {
               {{ amount | hdToUserValue | number: '2.2-2' }}
             } @else {
@@ -78,13 +66,23 @@ import { ToUserValuePipe } from '../shared';
       </div>
     </mat-card>
   `,
+  imports: [
+    DecimalPipe,
+    MatButton,
+    MatCard,
+    MatIcon,
+    CdkCopyToClipboard,
+    LetDirective,
+    HdObscureAddressPipe,
+    ToUserValuePipe,
+  ],
+  standalone: true,
 })
 export class PaymentSectionComponent {
-  @HostBinding() class = 'block';
-  @Input({ required: true }) amount: number | null = null;
-  @Input({ required: true }) memo: string | null = null;
-  @Input({ required: true }) requester: string | null = null;
-  @Output() approvePayment = new EventEmitter();
+  readonly amount = input.required<number | null>();
+  readonly memo = input.required<string | null>();
+  readonly requester = input.required<string | null>();
+  @Output() readonly approvePayment = new EventEmitter();
 
   onApprovePayment() {
     this.approvePayment.emit();
