@@ -1,8 +1,8 @@
-import { Component, EventEmitter, Input, Output, inject } from '@angular/core';
+import { Component, EventEmitter, Output, inject, input } from '@angular/core';
 import { FormsModule, NgForm } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
+import { MatButton } from '@angular/material/button';
+import { MatError, MatFormField, MatLabel } from '@angular/material/form-field';
+import { MatInput } from '@angular/material/input';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { fromUserValue } from '../utils';
 
@@ -25,7 +25,7 @@ export interface RequestPaymentFormPayload {
         <input
           matInput
           name="memo"
-          [(ngModel)]="model.memo"
+          [(ngModel)]="model().memo"
           #memoControl="ngModel"
           required
         />
@@ -44,7 +44,7 @@ export interface RequestPaymentFormPayload {
           matInput
           name="amount"
           type="number"
-          [(ngModel)]="model.amount"
+          [(ngModel)]="model().amount"
           #amountControl="ngModel"
           required
           min="0.01"
@@ -64,7 +64,7 @@ export interface RequestPaymentFormPayload {
       <footer>
         <button
           type="submit"
-          [disabled]="disabled"
+          [disabled]="disabled()"
           mat-raised-button
           color="primary"
         >
@@ -73,32 +73,31 @@ export interface RequestPaymentFormPayload {
       </footer>
     </form>
   `,
+  imports: [FormsModule, MatFormField, MatInput, MatLabel, MatError, MatButton],
   standalone: true,
-  imports: [FormsModule, MatFormFieldModule, MatInputModule, MatButtonModule],
 })
 export class RequestPaymentFormComponent {
   private readonly _matSnackBar = inject(MatSnackBar);
 
-  @Input() model: RequestPaymentFormModel = {
+  readonly model = input<RequestPaymentFormModel>({
     amount: null,
     memo: null,
-  };
-  @Input() disabled = false;
-  @Output() requestPayment = new EventEmitter<RequestPaymentFormPayload>();
+  });
+  readonly disabled = input(false);
+  @Output() readonly requestPayment =
+    new EventEmitter<RequestPaymentFormPayload>();
 
   onSubmit(form: NgForm) {
-    if (
-      form.invalid ||
-      this.model.amount === null ||
-      this.model.memo === null
-    ) {
+    const model = this.model();
+
+    if (form.invalid || model.amount === null || model.memo === null) {
       this._matSnackBar.open('Invalid data, review form entries.', 'close', {
         duration: 3000,
       });
     } else {
       this.requestPayment.emit({
-        amount: fromUserValue(this.model.amount),
-        memo: this.model.memo,
+        amount: fromUserValue(model.amount),
+        memo: model.memo,
       });
     }
   }
