@@ -1,51 +1,72 @@
-import { Component, Input } from '@angular/core';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { TransactionStatus } from '../utils';
+import { Component, EventEmitter, Output, input } from '@angular/core';
+import { MatButton } from '@angular/material/button';
+import { MatProgressSpinner } from '@angular/material/progress-spinner';
+import { LetDirective } from '@ngrx/component';
+import { TransactionState } from '../utils';
 
 @Component({
   selector: 'my-bank-process-transaction-section',
   template: `
-    @if (
-      status === 'pending' || status === 'sending' || status === 'confirming'
-    ) {
-      <div class="flex flex-col gap-4 items-center">
-        <mat-progress-spinner
-          mode="indeterminate"
-          diameter="64"
-        ></mat-progress-spinner>
-        <p class="text-sm italic text-center">
-          Process... Do NOT reload the page.
-        </p>
-      </div>
-    } @else if (status === 'confirmed') {
-      <div>
-        <p class="mb-2">
-          Feel free to inspect the transaction on the Solana Explorer:
-        </p>
-        <p>
-          <a
-            [href]="explorerUrl"
-            target="_blank"
-            class="underline text-blue-400"
+    <div
+      class="flex flex-col gap-4 items-center"
+      *ngrxLet="transactionState() as transactionState"
+    >
+      @switch (transactionState.status) {
+        @case ('pending') {
+          <p>
+            Your transaction is ready to be sent. Press the button to proceed.
+          </p>
+
+          <button
+            mat-raised-button
+            color="primary"
+            (click)="onSendTransaction()"
           >
-            [view in explorer]
-          </a>
-        </p>
-      </div>
-    } @else if (status === 'failed') {
-      <div>
-        <p class="px-4 py-1 bg-red-200 text-red-600">
-          {{ error }}
-        </p>
-      </div>
-    }
+            Send Transaction
+          </button>
+        }
+
+        @case ('confirmed') {
+          <p class="mb-2">
+            Feel free to inspect the transaction on the Solana Explorer:
+          </p>
+          <p>
+            <a
+              [href]="transactionState.explorerUrl"
+              target="_blank"
+              class="underline text-blue-400"
+            >
+              [view in explorer]
+            </a>
+          </p>
+        }
+
+        @case ('failed') {
+          <p class="px-4 py-1 bg-red-200 text-red-600">
+            {{ transactionState.error }}
+          </p>
+        }
+
+        @default {
+          <mat-progress-spinner
+            mode="indeterminate"
+            diameter="64"
+          ></mat-progress-spinner>
+          <p class="text-sm italic text-center">
+            Processing... Do NOT reload the page.
+          </p>
+        }
+      }
+    </div>
   `,
+  imports: [MatProgressSpinner, MatButton, LetDirective],
   standalone: true,
-  imports: [MatProgressSpinnerModule],
 })
 export class ProcessTransactionSectionComponent {
-  @Input({ required: true }) status: TransactionStatus = 'pending';
-  @Input({ required: true }) signature: string | null = null;
-  @Input({ required: true }) error: string | null = null;
-  @Input({ required: true }) explorerUrl: string | null = null;
+  readonly transactionState = input.required<TransactionState>();
+  @Output() readonly sendTransaction = new EventEmitter();
+
+  onSendTransaction() {
+    this.sendTransaction.emit();
+  }
 }
