@@ -1,8 +1,8 @@
-import { Component, EventEmitter, Input, Output, inject } from '@angular/core';
+import { Component, EventEmitter, Output, inject, input } from '@angular/core';
 import { FormsModule, NgForm } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
+import { MatButton } from '@angular/material/button';
+import { MatError, MatFormField, MatLabel } from '@angular/material/form-field';
+import { MatInput } from '@angular/material/input';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { fromUserValue } from '../utils';
 
@@ -27,7 +27,7 @@ export interface TransferFormPayload {
         <input
           matInput
           name="receiver"
-          [(ngModel)]="model.receiver"
+          [(ngModel)]="model().receiver"
           #receiverControl="ngModel"
           required
         />
@@ -45,7 +45,7 @@ export interface TransferFormPayload {
         <input
           matInput
           name="memo"
-          [(ngModel)]="model.memo"
+          [(ngModel)]="model().memo"
           #memoControl="ngModel"
           required
         />
@@ -64,7 +64,7 @@ export interface TransferFormPayload {
           matInput
           name="amount"
           type="number"
-          [(ngModel)]="model.amount"
+          [(ngModel)]="model().amount"
           #amountControl="ngModel"
           required
           min="0.01"
@@ -83,7 +83,7 @@ export interface TransferFormPayload {
       <footer>
         <button
           type="submit"
-          [disabled]="disabled"
+          [disabled]="disabled()"
           mat-raised-button
           color="primary"
         >
@@ -92,35 +92,37 @@ export interface TransferFormPayload {
       </footer>
     </form>
   `,
+  imports: [FormsModule, MatFormField, MatInput, MatLabel, MatError, MatButton],
   standalone: true,
-  imports: [FormsModule, MatFormFieldModule, MatInputModule, MatButtonModule],
 })
 export class TransferFormComponent {
   private readonly _matSnackBar = inject(MatSnackBar);
 
-  @Input() model: TransferFormModel = {
+  readonly model = input<TransferFormModel>({
     receiver: null,
     amount: null,
     memo: null,
-  };
-  @Input() disabled = false;
-  @Output() transfer = new EventEmitter<TransferFormPayload>();
+  });
+  readonly disabled = input(false);
+  @Output() readonly transfer = new EventEmitter<TransferFormPayload>();
 
   onSubmit(form: NgForm) {
+    const model = this.model();
+
     if (
       form.invalid ||
-      this.model.amount === null ||
-      this.model.receiver === null ||
-      this.model.memo === null
+      model.amount === null ||
+      model.receiver === null ||
+      model.memo === null
     ) {
       this._matSnackBar.open('Invalid data, review form entries.', 'close', {
         duration: 3000,
       });
     } else {
       this.transfer.emit({
-        amount: fromUserValue(this.model.amount),
-        receiver: this.model.receiver,
-        memo: this.model.memo,
+        amount: fromUserValue(model.amount),
+        receiver: model.receiver,
+        memo: model.memo,
       });
     }
   }
