@@ -1,22 +1,20 @@
 import { Component, OnInit, inject } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
+import { MatButton } from '@angular/material/button';
+import { MatCard } from '@angular/material/card';
+import { MatDialog } from '@angular/material/dialog';
+import { MatIcon } from '@angular/material/icon';
 import { RouterOutlet } from '@angular/router';
 import { ConnectionStore } from '@heavy-duty/wallet-adapter';
 import { HdWalletMultiButtonComponent } from '@heavy-duty/wallet-adapter-material';
-import { ShyftApiKeyService } from './core';
-import { UpdateSettingsService } from './settings';
+import { lastValueFrom } from 'rxjs';
+import { ShyftApiService } from './core';
+import {
+  UpdateSettingsFormPayload,
+  UpdateSettingsModalComponent,
+  UpdateSettingsModalData,
+} from './settings';
 
 @Component({
-  standalone: true,
-  imports: [
-    RouterOutlet,
-    MatButtonModule,
-    MatCardModule,
-    MatIconModule,
-    HdWalletMultiButtonComponent,
-  ],
   selector: 'my-bank-root',
   template: `
     <header class="pb-8 pt-16 relative">
@@ -37,21 +35,38 @@ import { UpdateSettingsService } from './settings';
       <router-outlet></router-outlet>
     </main>
   `,
+  imports: [
+    RouterOutlet,
+    MatButton,
+    MatCard,
+    MatIcon,
+    HdWalletMultiButtonComponent,
+  ],
+  standalone: true,
 })
 export class AppComponent implements OnInit {
   private readonly _connectionStore = inject(ConnectionStore);
-  private readonly _shyftApiKeyService = inject(ShyftApiKeyService);
-  private readonly _updateSettingsService = inject(UpdateSettingsService);
+  private readonly _shyftApiKeyService = inject(ShyftApiService);
+  private readonly _matDialog = inject(MatDialog);
 
   ngOnInit() {
     this._connectionStore.setEndpoint(this._shyftApiKeyService.endpoint());
   }
 
   async onUpdateSettings() {
-    const updateSettingsPayload =
-      await this._updateSettingsService.updateSettings({
-        shyftApiKey: this._shyftApiKeyService.shyftApiKey(),
-      });
+    const updateSettingsPayload = await lastValueFrom(
+      this._matDialog
+        .open<
+          UpdateSettingsModalComponent,
+          UpdateSettingsModalData,
+          UpdateSettingsFormPayload
+        >(UpdateSettingsModalComponent, {
+          data: {
+            shyftApiKey: this._shyftApiKeyService.shyftApiKey(),
+          },
+        })
+        .afterClosed(),
+    );
 
     if (updateSettingsPayload) {
       this._shyftApiKeyService.setShyftApiKey(
