@@ -1,3 +1,2 @@
-export * from './shyft-api-key.service';
-export * from './transaction-api.service';
+export * from './shyft-api.service';
 export * from './wallet.service';
