@@ -1,7 +1,7 @@
 import { Component, inject } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
+import { MatButton } from '@angular/material/button';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { MatIconModule } from '@angular/material/icon';
+import { MatIcon } from '@angular/material/icon';
 import { PublicKey } from '@solana/web3.js';
 import {
   config,
@@ -34,8 +34,8 @@ export interface PaymentRequestModalData {
       ></my-bank-payment-request-section>
     </div>
   `,
+  imports: [MatButton, MatIcon, PaymentRequestSectionComponent],
   standalone: true,
-  imports: [MatButtonModule, MatIconModule, PaymentRequestSectionComponent],
 })
 export class PaymentRequestModalComponent {
   private readonly _matDialogRef = inject(
