@@ -1,7 +1,7 @@
 import { Component, inject } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
+import { MatButton } from '@angular/material/button';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { MatIconModule } from '@angular/material/icon';
+import { MatIcon } from '@angular/material/icon';
 import {
   UpdateSettingsFormComponent,
   UpdateSettingsFormPayload,
@@ -37,8 +37,8 @@ export interface UpdateSettingsModalData {
       </a>
     </footer>
   `,
+  imports: [MatButton, MatIcon, UpdateSettingsFormComponent],
   standalone: true,
-  imports: [MatButtonModule, MatIconModule, UpdateSettingsFormComponent],
 })
 export class UpdateSettingsModalComponent {
   private readonly _matDialogRef = inject(
