@@ -1,7 +1,7 @@
 import { Component, inject } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
+import { MatButton } from '@angular/material/button';
 import { MatDialogRef } from '@angular/material/dialog';
-import { MatIconModule } from '@angular/material/icon';
+import { MatIcon } from '@angular/material/icon';
 import {
   TransferFormComponent,
   TransferFormPayload,
@@ -23,8 +23,8 @@ import {
       ></my-bank-transfer-form>
     </div>
   `,
+  imports: [MatButton, MatIcon, TransferFormComponent],
   standalone: true,
-  imports: [MatButtonModule, MatIconModule, TransferFormComponent],
 })
 export class TransferModalComponent {
   private readonly _matDialogRef = inject(MatDialogRef<TransferModalComponent>);
