@@ -32,6 +32,7 @@ export type TransactionState =
       status: 'confirmed';
       signature: TransactionSignature;
       error: null;
+      explorerUrl: string;
     }
   | {
       status: 'failed';
@@ -45,7 +46,7 @@ export type SendFunction = (
   connection: Connection,
   publicKey: PublicKey,
   transactionInstructions: TransactionInstruction[],
-) => Promise<void>;
+) => Promise<TransactionSignature>;
 
 export type TransactionSenderSignal = WritableSignal<TransactionState> & {
   send: SendFunction;
@@ -125,13 +126,17 @@ export function createTransactionSender(
           status: 'confirmed',
           signature,
           error: null,
+          explorerUrl: `https://explorer.solana.com/tx/${signature}`,
         });
+
+        return signature;
       } catch (error) {
         state.set({
           status: 'failed',
           signature: null,
           error: stringifyError(error),
         });
+        throw error;
       }
     },
   );
